requirements.txt
snowflake-connector-python>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0 